feat: optimize UDF argument handling and caching mechanisms
